# db: lazily construct combined iterator 