Combined range-key iteration adds overhead to Pebble iterators. Construction of
the range-key iterator stack adds a noticeable slowdown, even when the LSM
contains no range keys. Additionally, when the LSM contains range keys outside
of the span being iterated over, additional key comparisons are expected to add
an additional slowdown. In order to optimize these cases for the expected case
of rare range keys, this commit adds lazy combined iteration.

When an iterator is constructed with the IterKeyTypePointsAndRanges key type,
iterator construction examines the indexed batch (if any) and memtables to see
if they contain any range keys. If they don't, the iterator is initially
constructed with only a point iterator stack. A flag signals to lower-level
iterators that the user configured the iterator for combined iteration, but
combined iteration isn't yet being used.

When a levelIter observes a file containing range keys, the levelIter sets
state to signal that there might be range keys that need to be interleaved. An
interstitial internal iterator `lazyCombinedIter` observes the signal,
constructs the range-key iterator and reconfigures the iterator stack to use
combined iteration.

While this commit is nice performance win, it unfortunately introduces several
subtle complexities:
  1. When performing lazy combined iteration, the merging iterator must not
     alter seek keys according to range deletions until AFTER the level is
     already positioned according the seek key. This allows levelIters the
     opportunity to observe range keys between the user-provided seek key and
     the new modified seek key.
  2. When the merging iterator does seek a level to a tombstone's end key, it
     must set a new RelativeSeek seek flag, which causes the levelIter to scan
     all the fileMetadata between the original seek and the modified seek,
     rather than skipping straight to the fileMetadata matching the modified
     seek.
  3. Deciding how to seek the range-key iterator when switching to combined
     iteration is full of intricacies depending on the iterator operation
     and context.

The below benchmarks capture the performance improvement of avoiding the
range-key iterator construction when no range keys exist in the LSM. There
aren't yet suitable benchmarks for measuring the improvement in cases where
there exist range keys in the LSM, but the range keys are outside the keyspace
being iterated over.

```
name                                                    old time/op    new time/op    delta
MVCCScan_Pebble/rows=1/versions=1/valueSize=64-10         3.04µs ± 5%    2.62µs ± 5%  -13.69%  (p=0.000 n=9+10)
MVCCScan_Pebble/rows=100/versions=1/valueSize=64-10       21.2µs ± 1%    20.7µs ± 1%   -2.74%  (p=0.000 n=9+10)
MVCCScan_Pebble/rows=10000/versions=1/valueSize=64-10     1.41ms ± 1%    1.36ms ± 1%   -3.10%  (p=0.000 n=9+10)
MVCCGet_Pebble/batch=false/versions=1/valueSize=8-10      3.02µs ± 2%    2.58µs ± 0%  -14.84%  (p=0.000 n=9+10)
MVCCGet_Pebble/batch=false/versions=10/valueSize=8-10     3.48µs ± 2%    3.06µs ± 4%  -12.08%  (p=0.000 n=9+9)
MVCCGet_Pebble/batch=false/versions=100/valueSize=8-10    7.29µs ± 2%    6.93µs ±11%   -4.88%  (p=0.028 n=9+10)
MVCCGet_Pebble/batch=true/versions=1/valueSize=8-10       1.59µs ± 1%    1.57µs ± 4%     ~     (p=0.128 n=9+10)
MVCCGet_Pebble/batch=true/versions=10/valueSize=8-10      2.18µs ± 3%    2.18µs ± 6%     ~     (p=0.593 n=10+10)
MVCCGet_Pebble/batch=true/versions=100/valueSize=8-10     5.55µs ± 9%    5.17µs ± 1%   -6.95%  (p=0.000 n=10+9)

name                                                    old speed      new speed      delta
MVCCScan_Pebble/rows=1/versions=1/valueSize=64-10       21.1MB/s ± 5%  24.4MB/s ± 5%  +15.90%  (p=0.000 n=9+10)
MVCCScan_Pebble/rows=100/versions=1/valueSize=64-10      301MB/s ± 1%   310MB/s ± 1%   +2.82%  (p=0.000 n=9+10)
MVCCScan_Pebble/rows=10000/versions=1/valueSize=64-10    454MB/s ± 1%   469MB/s ± 1%   +3.20%  (p=0.000 n=9+10)
MVCCGet_Pebble/batch=false/versions=1/valueSize=8-10    2.65MB/s ± 2%  3.11MB/s ± 1%  +17.44%  (p=0.000 n=9+10)
MVCCGet_Pebble/batch=false/versions=10/valueSize=8-10   2.30MB/s ± 2%  2.62MB/s ± 4%  +13.86%  (p=0.000 n=9+9)
MVCCGet_Pebble/batch=false/versions=100/valueSize=8-10  1.10MB/s ± 1%  1.16MB/s ±10%   +5.66%  (p=0.031 n=8+10)
MVCCGet_Pebble/batch=true/versions=1/valueSize=8-10     5.04MB/s ± 1%  5.10MB/s ± 4%     ~     (p=0.127 n=9+10)
MVCCGet_Pebble/batch=true/versions=10/valueSize=8-10    3.68MB/s ± 3%  3.68MB/s ± 6%     ~     (p=0.589 n=10+10)
MVCCGet_Pebble/batch=true/versions=100/valueSize=8-10   1.44MB/s ± 9%  1.54MB/s ± 2%   +7.00%  (p=0.000 n=10+10)
```

Close https://github.com/cockroachdb/pebble/issues/1780.