@erikgrinaker 2022-06-29
Love the speedups, but share your hesitation about the complexity. Is it possible to constrain this e.g. by introducing new abstraction layers?

We also need to be careful that we don't fall off a performance cliff when range keys do show up (initially due to import rollbacks). We haven't done any benchmarking of this case yet, but my sense is that it's going to be pretty severe at current, so we're likely going to have to do a substantial amount of optimization for this case as well. Are there other alternatives to make combined iterator construction cheaper? Do you think it's worthwhile to eat this complexity cost regardless, or should we explore other options first?