@jbowens 2022-06-29
The iterator construction code path is already very minimal, at least in the case with no range keys in the LSM.

I haven't been able to think of any viable alternatives—I experimented with an alternative that cached `Iterator`s when `Close`d so that they can be reused on `NewIter`. That approach brought its own host of complexities and had a lessor (~5%) impact on the benchmarks. It could probably be optimized further, maybe with lock-free synchronization, but it brings its own challenges. I also think the caching approach would become less effective with higher compaction concurrency settings, at least without additional complexity of its own.

I'm leaning towards eating the complexity regardless, but I would like to spend more time trying to pare down its complexity, especially after spending some time away from the problem.

I think the biggest reason to go this route is that I expect it will help significantly in the case that the LSM contains range keys, but those range keys are in a part of the keyspace untouched by the iterator. Currently, I expect each LSM level containing a range key will have a significant impact on performance: It'll introduce key comparisons per-level per-seek. There are probably optimizations that can help mitigate this impact on repeated seeks in the same part of the keyspace, but even then the impact on the first seek will be significant.