@erikgrinaker 2022-06-29
Makes sense. We could possibly be more aggressive about iterator reuse in CRDB too, could be worth pooling and reconfiguring them rather than cloning -- at least on the read path. Will give it a spin when I get back to optimization work.