@erikgrinaker 2022-07-21
>Does this change cause any performance improvement for tpcc or kv95?

I did a couple of quick three-minute `kv95/enc=false/nodes=3/cpu=32` runs with this against `master`, showed a ~3% improvement. That recovers about half of the 7% regression we have since 22.1. I agree that the complexity is unfortunate, but at the same time, we need to claw back most of those 7%.